requirements.txt
python-telegram-bot==21.10
groq==0.18.0
python-dotenv==1.0.1
pandas==2.2.3
openpyxl==3.1.5
APScheduler==3.11.0
